requirements.txt
python-dotenv==1.0.1
openai==1.43.0
tenacity==9.0.0
loguru==0.7.2
playwright==1.47.0
requests==2.32.3
httpx<0.28
beautifulsoup4==4.12.3
lxml==5.3.0
tweepy==4.14.0
atproto==0.0.56

